requirements.txt
pandas==2.2.2
SQLAlchemy==2.0.32
psycopg2-binary==2.9.6
python-dotenv==1.0.1